backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "emoji-country-flag>=2.0.1",
    "fastapi>=0.115.12",
    "langcodes>=3.5.0",
    "langdetect>=1.0.9",
    "openai>=1.75.0",
    "pycountry>=24.6.1",
    "pydantic>=2.11.2",
    "uvicorn>=0.34.0",
]
